Implement durable functions
