+++ src/ScoopSearch.Functions/Constants.cs
@@ -2,7 +2,6 @@
 {
     public static class Constants
     {
-        public const string BucketsQueue = "scoopsearch-buckets";
         public const string GitHubHttpClientName = "GitHub";
         public const string GitHubHttpClientNoRedirectName = "GitHubNoRedirect";
     }
