--- src/ScoopSearch.Functions/Data/QueueItem.cs
@@ -1,20 +1,0 @@
-﻿using System;
-
-namespace ScoopSearch.Functions.Data
-{
-    public class QueueItem
-    {
-        public QueueItem(Uri bucket, int stars, bool official)
-        {
-            Bucket = bucket;
-            Stars = stars;
-            Official = official;
-        }
-
-        public Uri Bucket { get; }
-
-        public int Stars { get; }
-
-        public bool Official { get; }
-    }
-}
