+++ src/ScoopSearch.Functions/Function/Version.cs
@@ -6,20 +6,19 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Extensions.Logging;
 
-namespace ScoopSearch.Functions.Function
-{
-    public class Version
-    {
-        [FunctionName("Version")]
-        public IActionResult Run(
-            [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
-            HttpRequest request,
-            ILogger log)
-        {
-            var assembly = GetType().Assembly;
-            var assemblyLocation = assembly.Location;
+namespace ScoopSearch.Functions.Function;
 
-            return new OkObjectResult($"{assembly.GetName().Name} - {FileVersionInfo.GetVersionInfo(assemblyLocation).ProductVersion} - {File.GetLastWriteTimeUtc(assemblyLocation):R}");
-        }
+public class Version
+{
+    [FunctionName("Version")]
+    public IActionResult Run(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
+        HttpRequest request,
+        ILogger log)
+    {
+        var assembly = GetType().Assembly;
+        var assemblyLocation = assembly.Location;
+
+        return new OkObjectResult($"{assembly.GetName().Name} - {FileVersionInfo.GetVersionInfo(assemblyLocation).ProductVersion} - {File.GetLastWriteTimeUtc(assemblyLocation):R}");
     }
 }
