+++ src/ScoopSearch.Functions/host.json
@@ -6,7 +6,9 @@
             "default": "Warning",
             "Host.Results": "Information", // Required to get proper monitor on Azure
             "Host.Aggregator": "Information", // Required to get proper monitor on Azure
-            "Function": "Information"
+            "Function": "Information",
+            "DurableTask.AzureStorage": "Warning",
+            "DurableTask.Core": "Warning"
         },
         "applicationInsights": {
             "samplingSettings": {
@@ -20,10 +22,8 @@
     },
     "functionTimeout": "00:09:59",
     "extensions": {
-        "queues": {
-            "maxPollingInterval": "00:00:15",
-            "maxDequeueCount": 2,
-            "batchSize": 1
+        "durableTask": {
+            "hubName": "%ScoopSearchDurableHubName%"
         }
     }
 }
