+++ src/ScoopSearch.Functions/local.settings.json-sample
@@ -4,7 +4,7 @@
         "FUNCTIONS_WORKER_RUNTIME": "dotnet",
 
         "AzureWebJobsStorage": "DefaultEndpointsProtocol=https;AccountName=scoopsearch;AccountKey=[ACCOUNTKEY];EndpointSuffix=core.windows.net",
-        "DispatchBucketsCrawlerCron": "0 5 */6 * * *",
+        "KickoffStartOrchestratorCron": "0 5 */6 * * *",
 
         "AzureSearchServiceUrl": "https://[SERVICENAME].search.windows.net",
         "AzureSearchAdminApiKey": "[ADMINAPIKEY]",
